Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
 int main() {
-    float a;
+    double a;
     scanf("%f", &a);
     if (a==12345.6787){
         printf("12345.6789");
     }
     else{
-        printf("You entered: %.4f",a);
+        printf("You entered: %.4lf",a);
     }
     return 0;
     }
